+++ great_expectations/datasource/data_connector/configured_asset_azure_data_connector.py
@@ -3,18 +3,13 @@ import os
 import re
 from typing import List, Optional
 
-try:
-    from azure.storage.blob import BlobServiceClient
-except ImportError:
-    BlobServiceClient = None
-
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.batch import BatchDefinition
 from great_expectations.core.batch_spec import AzureBatchSpec, PathBatchSpec
-from great_expectations.datasource.data_connector import (
+from great_expectations.datasource.data_connector.asset import Asset
+from great_expectations.datasource.data_connector.configured_asset_file_path_data_connector import (
     ConfiguredAssetFilePathDataConnector,
 )
-from great_expectations.datasource.data_connector.asset import Asset
 from great_expectations.datasource.data_connector.util import list_azure_keys
 from great_expectations.execution_engine import (
     ExecutionEngine,
@@ -23,6 +18,14 @@ from great_expectations.execution_engine import (
 )
 
 logger = logging.getLogger(__name__)
+
+try:
+    from azure.storage.blob import BlobServiceClient
+except ImportError:
+    BlobServiceClient = None
+    logger.debug(
+        "Unable to load BlobServiceClient connection object; install optional Azure Storage Blob dependency for support"
+    )
 
 
 class ConfiguredAssetAzureDataConnector(ConfiguredAssetFilePathDataConnector):
@@ -161,7 +164,7 @@ class ConfiguredAssetAzureDataConnector(ConfiguredAssetFilePathDataConnector):
     ) -> str:
         # asset isn't used in this method.
         # It's only kept for compatibility with parent methods.
-        # Pandas and Spark execution engines require separate paths for compatibility with Azure's API.
+        # Pandas and Spark execution engines utilize separate path formats for accessing Azure Blob Storage service.
         full_path: str
         if isinstance(self.execution_engine, PandasExecutionEngine):
             full_path = os.path.join(